+++ test/fixtures/updater/language-tour/sources/basic.dart
@@ -1,0 +1,4 @@
+// #docregion greeting
+var greeting = 'hello';
+var scope = 'world';
+// #enddocregion greeting