+++ test/fixtures/updater/plaster-defaults/sources/plaster.dart
@@ -1,0 +1,13 @@
+// #docregion
+var greeting = 'hello';
+// #enddocregion
+// #docregion
+var scope = 'world';
+// #enddocregion
+
+// #docregion html
+<p>
+// #enddocregion
+// #docregion html
+</p>
+// #enddocregion